+++ samples/android/src/main/kotlin/com/maximbircu/devtools/configscreens/JsonSchemaSourceConfigActivity.kt
@@ -3,6 +3,8 @@ package com.maximbircu.devtools.configscreens
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import android.view.Gravity
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.maximbircu.devtools.R.layout
 import com.maximbircu.devtools.android.DevToolsConfigurationScreen
@@ -11,6 +13,7 @@ import com.maximbircu.devtools.common.readers.DevToolsSources
 import com.maximbircu.devtools.common.readers.json
 import kotlinx.android.synthetic.main.activity_tools_container.devToolsContainer
 
+@Suppress("LongMethod")
 class JsonSchemaSourceConfigActivity : AppCompatActivity() {
     private lateinit var devtools: DevTools
 
@@ -27,6 +30,16 @@ class JsonSchemaSourceConfigActivity : AppCompatActivity() {
 
         val source = DevToolsSources.json(assets.open("dev-tools.json").reader().readText())
         devtools = DevTools.create(source)
+
+        devtools.onConfigUpdated = { isCriticalUpdate ->
+            val toast = Toast.makeText(
+                this,
+                "A ${if (isCriticalUpdate) "critical" else "non-critical"} updated performed!",
+                Toast.LENGTH_SHORT
+            )
+            toast.setGravity(Gravity.CENTER, 0, 0)
+            toast.show()
+        }
 
         DevToolsConfigurationScreen.attachToView(devToolsContainer, devtools)
     }